% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{aq_documentation}
\alias{aq_documentation}
\title{Documentation}
\usage{
aq_documentation(path_append, domain = aq_domain())
}
\arguments{
\item{path_append}{A string to append to the URL path.}

\item{domain}{A string of the domain name.}
}
\description{
Documentation
}
\examples{
 aq_documentation("GetAuthTokenServiceRequest")
}
