% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{aq_domain}
\alias{aq_domain}
\title{Domain}
\usage{
aq_domain(domain = getOption("aq.domain", Sys.getenv("AQ_DOMAIN")))
}
\arguments{
\item{domain}{A string of the domain name.}
}
\description{
Domain
}
\examples{
aq_domain()
}
