% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-auth-token.R
\name{aq_get_auth_token}
\alias{aq_get_auth_token}
\alias{aq_token}
\title{Get Authorization Token}
\usage{
aq_get_auth_token(
  username = getOption("aq.username", Sys.getenv("AQ_USERNAME")),
  password = getOption("aq.password", Sys.getenv("AQ_PASSWORD")),
  domain = aq_domain()
)

aq_token(
  username = getOption("aq.username", Sys.getenv("AQ_USERNAME")),
  password = getOption("aq.password", Sys.getenv("AQ_PASSWORD"))
)
}
\arguments{
\item{username}{A string of the user name.}

\item{password}{A string of the user's password.}

\item{domain}{A string of the domain name.}
}
\value{
A character string of the temporary authorization token
}
\description{
Generates an temporary authorization token from the Aquarius API for a supplied
username and password.
}
\details{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=GetAuthTokenServiceRequest}
}
\section{Functions}{
\itemize{
\item \code{aq_token()}: Get Temporary Authorization Token

}}
\examples{
\dontrun{
  aq_get_auth_token()
}
}
