% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-location-data.R
\name{aq_get_location_data}
\alias{aq_get_location_data}
\title{Get Tables for Aquarius Database}
\usage{
aq_get_location_data(location, token = aq_token())
}
\arguments{
\item{location}{A string of the location name.}

\item{token}{A string of the user's authentication token.}
}
\value{
A tibble with xx
}
\description{
Get Tables for Aquarius Database
}
\examples{
\dontrun{
 aq_get_location_data("LC_DRY_WQ06_TEMP")
}
}
