% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-location-notes.R
\name{aq_get_location_notes}
\alias{aq_get_location_notes}
\title{Get Location Notes}
\usage{
aq_get_location_notes(
  location_identifier,
  ...,
  token = aq_token(),
  domain = aq_domain()
)
}
\arguments{
\item{location_identifier}{A string of the location identifier.}

\item{...}{These dots are for future extensions and must be empty.}

\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Location Notes
}
\examples{
\dontrun{
aq_get_location_notes("LC_DRY_WQ06_TEMP")
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=LocationNotesServiceRequest}
}
