% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-parameter-list.R
\name{aq_get_parameter_list}
\alias{aq_get_parameter_list}
\title{Get Parameter List}
\usage{
aq_get_parameter_list(token = aq_token(), domain = aq_domain())
}
\arguments{
\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Parameter List
}
\examples{
\dontrun{
 aq_get_parameter_list()
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=ParameterListServiceRequest}
}
