% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-series-corrected-data.R
\name{aq_get_time_series_corrected_data}
\alias{aq_get_time_series_corrected_data}
\title{Get Time Series Corrected Data}
\usage{
aq_get_time_series_corrected_data(
  time_series_id,
  query_from = NULL,
  query_to = NULL,
  ...,
  token = aq_token(),
  domain = aq_domain()
)
}
\arguments{
\item{time_series_id}{A string of time series ID.}

\item{query_from}{A date of the start of the period of interest.}

\item{query_to}{A date of the end of the period of interest.}

\item{...}{These dots are for future extensions and must be empty.}

\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Time Series Corrected Data
}
\examples{
\dontrun{
aq_get_time_series_corrected_data("27a6b0badd044e0c9b29d589b9e078d9")
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=TimeSeriesDataCorrectedServiceRequest}
}
