% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-series-data-corrected.R
\name{aq_get_time_series_data_corrected}
\alias{aq_get_time_series_data_corrected}
\title{Get Location Data}
\usage{
aq_get_time_series_data_corrected(
  time_series_id,
  ...,
  token = aq_token(),
  domain = aq_domain()
)
}
\arguments{
\item{time_series_id}{A string of time series ID.}

\item{...}{These dots are for future extensions and must be empty.}

\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Location Data
}
\examples{
\dontrun{
 aq_get_time_series_data_corrected("27a6b0badd044e0c9b29d589b9e078d9")
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=TimeSeriesDataCorrectedServiceRequest}
}
