% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-series-description-list-by-unique-id.R
\name{aq_get_time_series_description_list_by_unique_id}
\alias{aq_get_time_series_description_list_by_unique_id}
\title{Get Time Series Description}
\usage{
aq_get_time_series_description_list_by_unique_id(
  time_series_unique_ids,
  ...,
  token = aq_token(),
  domain = aq_domain()
)
}
\arguments{
\item{time_series_unique_ids}{A character vector of the time series ID.}

\item{...}{These dots are for future extensions and must be empty.}

\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Time Series Description
}
\examples{
\dontrun{
aq_get_time_series_description_list_by_unique_id("27a6b0badd044e0c9b29d589b9e078d9")
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=TimeSeriesDescriptionListByUniqueIdServiceRequest}
}
