% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-unit-list.R
\name{aq_get_unit_list}
\alias{aq_get_unit_list}
\title{Get Tables for Aquarius Database}
\usage{
aq_get_unit_list(token = aq_token())
}
\arguments{
\item{token}{A string of the user's authentication token.}
}
\description{
Get Tables for Aquarius Database
}
\examples{
\dontrun{
 aq_get_unit_list()
}
}
