% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{aq_tables}
\alias{aq_tables}
\title{Get Tables for Aquarius Database}
\usage{
aq_tables(token = aq_token())
}
\arguments{
\item{token}{A string of the user's authentication token.}
}
\value{
A tibble with xx
}
\description{
Get Tables for Aquarius Database
}
\examples{
\dontrun{
 aq_tables()
}
}
