% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{aq_token}
\alias{aq_token}
\title{Get Aquarius Token}
\usage{
aq_token(
  username = getOption("aq.username", Sys.getenv("AQ_USERNAME")),
  password = getOption("aq.password", Sys.getenv("AQ_PASSWORD"))
)
}
\arguments{
\item{username}{A string of the user name.}

\item{password}{A string of the user's password.}
}
\value{
A character string of the temporary access token
}
\description{
Generates an temporary authentication token from the Aquarius API for a supplied
username and password.
}
\examples{
\dontrun{
  aq_token()
}
}
