% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{aq_url}
\alias{aq_url}
\title{URL}
\usage{
aq_url(domain = aq_domain())
}
\arguments{
\item{domain}{A string of the domain name.}
}
\description{
URL
}
\examples{
aq_url("poissonconsulting.aquaticinformatics.net")
}
