% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquariusapi-package.R
\docType{package}
\name{aquariusapi-package}
\alias{aquariusapi}
\alias{aquariusapi-package}
\title{aquariusapi: An API to Aquarius Databases}
\description{
Provides an low-level interface to the Aquarius API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/aquariusapi/}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@pissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Sarah Lyons \email{sarah@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-6745-6796}{ORCID})
}

Other contributors:
\itemize{
  \item Elk Valley Resources [funder, copyright holder]
}

}
\keyword{internal}
