% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-location-data.R
\name{aq_get_location_data}
\alias{aq_get_location_data}
\title{Get Location Data}
\usage{
aq_get_location_data(
  location_identifier,
  ...,
  token = aq_token(),
  domain = aq_domain()
)
}
\arguments{
\item{location_identifier}{A string of the location identifier.}

\item{...}{These dots are for future extensions and must be empty.}

\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Location Data
}
\examples{
\dontrun{
aq_get_location_data("LC_DRY_WQ06_TEMP")
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=LocationDataServiceRequest}
}
