% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-location-data.R
\name{aq_get_location_data}
\alias{aq_get_location_data}
\title{Get Location Data}
\usage{
aq_get_location_data(location, token = aq_token())
}
\arguments{
\item{location}{A string of the location name.}

\item{token}{A string of the user's authentication token.}
}
\description{
Get Location Data
}
\examples{
\dontrun{
 aq_get_location_data("LC_DRY_WQ06_TEMP")
}
}
