% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-parameter-list.R
\name{aq_get_parameter_list}
\alias{aq_get_parameter_list}
\title{Get Tables for Aquarius Database}
\usage{
aq_get_parameter_list(token = aq_token())
}
\arguments{
\item{token}{A string of the user's authentication token.}
}
\description{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=ParameterListServiceRequest}
}
\examples{
\dontrun{
 aq_get_parameter_list()
}
}
