% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-series-raw-data.R
\name{aq_get_time_series_raw_data}
\alias{aq_get_time_series_raw_data}
\title{Get Location Data}
\usage{
aq_get_time_series_raw_data(time_series_id, token = aq_token())
}
\arguments{
\item{time_series_id}{A string of time series ID.}

\item{token}{A string of the user's authentication token.}
}
\description{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=GetTimeSeriesRawData}
}
\examples{
\dontrun{
 aq_get_time_series_raw_data("27a6b0badd044e0c9b29d589b9e078d9")
}
}
