% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-time-series-unique-id-list.R
\name{aq_get_time_series_unique_id_list}
\alias{aq_get_time_series_unique_id_list}
\title{Get Time Series Unique ID List}
\usage{
aq_get_time_series_unique_id_list(token = aq_token(), domain = aq_domain())
}
\arguments{
\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=TimeSeriesUniqueIdListServiceRequest}
}
\examples{
\dontrun{
 aq_get_time_series_unique_id_list()
}
}
