% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-unit-list.R
\name{aq_get_unit_list}
\alias{aq_get_unit_list}
\title{Get Unit List}
\usage{
aq_get_unit_list(..., token = aq_token(), domain = aq_domain())
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{token}{A string of the user's authentication token.}

\item{domain}{A string of the domain name.}
}
\description{
Get Unit List
}
\examples{
\dontrun{
aq_get_unit_list()
}
}
\seealso{
\url{https://evr.aquaticinformatics.net/AQUARIUS/Publish/v2/json/metadata?op=UnitListServiceRequest}
}
