% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions for aquarius2r Package}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{domain}{A string of the domain name.}

\item{location_identifier}{A string of the location identifier.}

\item{parameter}{A string of the parameter.}

\item{password}{A string of the user's password.}

\item{path_append}{A string to append to the URL path.}

\item{time_series_id}{A string of time series ID.}

\item{time_series_ids}{A character vector of the time series ID.}

\item{token}{A string of the user's authentication token.}

\item{query_from}{A date of the start of the period of interest.}

\item{query_to}{A date of the end of the period of interest.}

\item{username}{A string of the user name.}
}
\value{
A tibble.
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.

A date is a non-missing character scalar.
}
\keyword{internal}
